backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
torch>=2.1.0
numpy>=1.24.3
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
transformers>=4.35.0
scikit-learn>=1.3.0
qdrant-client>=1.9.0
openai>=1.0.0
